Separate MachineList tests to relevant components.
